Split agentic_process builtin into package; add AgenticProcessStatus.NEW
