+++ flow_sdk/builtin/agentic_process/__init__.py
@@ -1,0 +1,51 @@
+"""Agentic process package — re-exports for backward compatibility.
+
+All imports from ``flow_sdk.builtin.agentic_process`` continue to work as before.
+"""
+
+from flow_sdk.fs_records.agentic_process_record import AgenticProcessStatus
+
+from flow_sdk.builtin.agentic_process._shared import (
+    AgenticContext,
+    ContextData,
+    ControlAppendRequest,
+    ControlContinueRequest,
+    ControlInputRequest,
+    ControlStartRequest,
+    ControlStepRequest,
+    CreateProcessRequest,
+    DebugState,
+    ExecuteRequest,
+    ProcessorState,
+    ProcessResultRequest,
+    RunFileRequest,
+    RunRequest,
+    StackFrame,
+    _send_flow_data_message,
+)
+from flow_sdk.builtin.agentic_process.agentic_process import AgenticProcess
+from flow_sdk.builtin.agentic_process.agentic_processor import AgenticProcessor
+from flow_sdk.builtin.agentic_process.apu import APU
+
+__all__ = [
+    "AgenticProcessStatus",
+    "StackFrame",
+    "DebugState",
+    "ProcessorState",
+    "AgenticContext",
+    "AgenticProcess",
+    "AgenticProcessor",
+    "APU",
+    "ContextData",
+    "ControlStartRequest",
+    "ControlInputRequest",
+    "ControlStepRequest",
+    "ControlAppendRequest",
+    "ControlContinueRequest",
+    "RunFileRequest",
+    "RunRequest",
+    "ExecuteRequest",
+    "CreateProcessRequest",
+    "ProcessResultRequest",
+    "_send_flow_data_message",
+]
